# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SystemStatus {
  isAppleSilicon: boolean;
  /** True when real MLX training can run on this machine */
  trainingBackendReady: boolean;
  /** True when running somewhere other than the user's Mac; downloads/training are simulated for preview purposes */
  simulationMode: boolean;
  platform: string;
  freeDiskGb: number;
  message: string;
}

export type ModelSize = typeof ModelSize[keyof typeof ModelSize];


export const ModelSize = {
  small: 'small',
  medium: 'medium',
  large: 'large',
} as const;

export type ModelStatus = typeof ModelStatus[keyof typeof ModelStatus];


export const ModelStatus = {
  not_downloaded: 'not_downloaded',
  downloading: 'downloading',
  ready: 'ready',
  failed: 'failed',
} as const;

/**
 * Core architecture family. Non-transformer types are "fast weights" architectures that keep a fixed-size internal memory state instead of a growing KV cache (state space models, linear attention / RWKV, and attention+SSM hybrids).
 */
export type ModelArchitecture = typeof ModelArchitecture[keyof typeof ModelArchitecture];


export const ModelArchitecture = {
  transformer: 'transformer',
  ssm: 'ssm',
  'linear-attention': 'linear-attention',
  hybrid: 'hybrid',
} as const;

/**
 * Whether LoRA fine-tuning of this model is a well-trodden path (supported), possible but less battle-tested (experimental), or not available (none).
 */
export type FineTuneSupport = typeof FineTuneSupport[keyof typeof FineTuneSupport];


export const FineTuneSupport = {
  supported: 'supported',
  experimental: 'experimental',
  none: 'none',
} as const;

export type ExportFormat = typeof ExportFormat[keyof typeof ExportFormat];


export const ExportFormat = {
  ollama: 'ollama',
  gguf: 'gguf',
} as const;

export interface Model {
  id: string;
  name: string;
  family: string;
  /** Human readable parameter count, e.g. "3B" */
  parameterCount: string;
  sizeCategory: ModelSize;
  sizeGb: number;
  description: string;
  recommendedUse: string;
  /** Plain language guidance, e.g. "Comfortable on any modern Mac" */
  memoryGuidance: string;
  architecture: ModelArchitecture;
  fineTuneSupport: FineTuneSupport;
  /** Export formats this model's architecture supports; empty when GGUF/Ollama packaging is unavailable. */
  exportFormats: ExportFormat[];
  status: ModelStatus;
  /** 0 to 100 */
  downloadProgress: number;
  /** @nullable */
  error?: string | null;
}

export interface DatasetUpload {
  file: Blob;
  name?: string;
}

export type DatasetStatus = typeof DatasetStatus[keyof typeof DatasetStatus];


export const DatasetStatus = {
  validating: 'validating',
  ready: 'ready',
  invalid: 'invalid',
} as const;

export interface DatasetRow {
  prompt: string;
  response: string;
}

export interface Dataset {
  id: string;
  name: string;
  status: DatasetStatus;
  rowCount: number;
  sizeBytes: number;
  createdAt: string;
  preview: DatasetRow[];
  /** @nullable */
  error?: string | null;
}

export interface TrainingPreset {
  id: string;
  name: string;
  description: string;
  /** Plain language estimate, e.g. "A few minutes" */
  estimatedTime: string;
  epochs: number;
  learningRate: number;
  loraRank: number;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  queued: 'queued',
  preparing: 'preparing',
  training: 'training',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
  exporting: 'exporting',
  exported: 'exported',
} as const;

export interface TrainingJobInput {
  name: string;
  modelId: string;
  datasetId: string;
  presetId: string;
}

export interface TrainingJob {
  id: string;
  name: string;
  modelId: string;
  modelName: string;
  datasetId: string;
  datasetName: string;
  presetId: string;
  presetName: string;
  status: JobStatus;
  /** 0 to 100 */
  progress: number;
  currentEpoch: number;
  totalEpochs: number;
  /** @nullable */
  loss: number | null;
  /** @nullable */
  etaSeconds: number | null;
  /** Human readable current status, no jargon or stack traces */
  statusMessage: string;
  /** @nullable */
  error?: string | null;
  /** True if this job ran in simulation mode rather than real MLX training */
  simulated: boolean;
  createdAt: string;
  exportReady: boolean;
  /** @nullable */
  exportFormat?: string | null;
  /** Recent human-readable log lines for this job (no raw stack traces) */
  logs?: string[];
  /** Loss value recorded at each training step, oldest first, for charting */
  lossHistory?: number[];
}

export type ExportInputFormat = typeof ExportInputFormat[keyof typeof ExportInputFormat];


export const ExportInputFormat = {
  ollama: 'ollama',
  gguf: 'gguf',
} as const;

export interface ExportInput {
  format: ExportInputFormat;
}

export type ChatRole = typeof ChatRole[keyof typeof ChatRole];


export const ChatRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatRole;
  content: string;
  timestamp: string;
}

export interface ChatSessionInput {
  modelId: string;
  /**
     * Optional completed fine-tuning job whose adapter should be applied on top of the base model
     * @nullable
     */
  jobId?: string | null;
}

export interface ChatMessageInput {
  content: string;
}

export interface ChatSession {
  id: string;
  /** Human readable session title, e.g. first user message snippet */
  title: string;
  modelId: string;
  modelName: string;
  /** @nullable */
  jobId?: string | null;
  /**
     * Name of the fine-tuning job whose adapter is applied, if any
     * @nullable
     */
  jobName?: string | null;
  /** True when replies are simulated because this isn't running on the user's Mac */
  simulated: boolean;
  /** True while the assistant reply is still streaming */
  generating: boolean;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
  /** Full message history (empty in list responses) */
  messages: ChatMessage[];
  /** @nullable */
  error?: string | null;
}

export interface TranscriptSelection {
  sessionId: string;
  /** Zero-based indices of the kept user/assistant exchange pairs within the session */
  exchangeIndices: number[];
}

export interface DatasetFromTranscriptsInput {
  name: string;
  /** When true, each example is framed as one of the model's own remembered experiences */
  memoryFraming: boolean;
  selections: TranscriptSelection[];
}

export interface ErrorResponse {
  error: string;
  suggestion?: string;
}