# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models/{modelId}:
    get:
      operationId: getModel
      tags: [models]
      summary: Get a single model's detail
      parameters:
        - $ref: "#/components/parameters/ModelIdParam"
      responses:
        "200":
          description: Model detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{modelId}/download:
    post:
      operationId: startModelDownload
      tags: [models]
      summary: Start downloading a model
      parameters:
        - $ref: "#/components/parameters/ModelIdParam"
      responses:
        "202":
          description: Download started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{modelId}/download/events:
    get:
      operationId: streamModelDownloadEvents
      tags: [models]
      summary: Server-sent events stream of download progress
      description: text/event-stream of ModelDownloadEvent payloads
      parameters:
        - $ref: "#/components/parameters/ModelIdParam"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  parameters:
    ModelIdParam:
      name: modelId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ModelSize:
      type: string
      enum: [small, medium, large]
    ModelStatus:
      type: string
      enum: [not_downloaded, downloading, ready, failed]
    ModelArchitecture:
      type: string
      enum: [transformer, ssm, linear-attention, hybrid]
      description: >
        Core architecture family. Non-transformer types are "fast weights"
        architectures that keep a fixed-size internal memory state instead of
        a growing KV cache (state space models, linear attention / RWKV, and
        attention+SSM hybrids).
    FineTuneSupport:
      type: string
      enum: [supported, experimental, none]
      description: >
        Whether LoRA fine-tuning of this model is a well-trodden path
        (supported), possible but less battle-tested (experimental), or not
        available (none).
    ExportFormat:
      type: string
      enum: [ollama, gguf]
    Model:
      type: object
      required:
        - id
        - name
        - family
        - parameterCount
        - sizeCategory
        - sizeGb
        - description
        - recommendedUse
        - memoryGuidance
        - architecture
        - fineTuneSupport
        - exportFormats
        - status
        - downloadProgress
      properties:
        id:
          type: string
        name:
          type: string
        family:
          type: string
        parameterCount:
          type: string
          description: Human readable parameter count, e.g. "3B"
        sizeCategory:
          $ref: "#/components/schemas/ModelSize"
        sizeGb:
          type: number
        description:
          type: string
        recommendedUse:
          type: string
        memoryGuidance:
          type: string
          description: Plain language guidance, e.g. "Comfortable on any modern Mac"
        architecture:
          $ref: "#/components/schemas/ModelArchitecture"
        fineTuneSupport:
          $ref: "#/components/schemas/FineTuneSupport"
        exportFormats:
          type: array
          items:
            $ref: "#/components/schemas/ExportFormat"
          description: Export formats this model's architecture supports; empty when GGUF/Ollama packaging is unavailable.
        status:
          $ref: "#/components/schemas/ModelStatus"
        downloadProgress:
          type: number
          description: 0 to 100
        error:
          type: ["string", "null"]
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        suggestion:
          type: string